requirements.txt
Flask >= 3.0.3
grpcio_tools
grpcio